Backend/requirements.txt
Django # BURADA OLMALI
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
djangorestframework-simplejwt
requests
django-cors-headers
pytesseract
Pillow